add image builder framework
