+++ template/src/aws_central_infrastructure/artifact_stores/lib/__init__.py
@@ -1,1 +1,3 @@
 from .code_artifact import RepoPackageClaims
+from .image_builder import ImageBuilderConfig
+from .image_builder import ImageShareConfig
